Control.Reducible.Lifting

Documentation

applyLifting :: MonadLifting m n => n a -> Lifting m n a

This function plays the role of lift for Lifting m.

The reason we don't use lift proper is that the exact implementation depends on both parameters m and n, but the kind of MonadTrans does not allow us that choice.

newtype Lifting (m :: Type -> Type) (n :: Type -> Type) a

A wrapper for decomposing monad transformer applications.

Lifting m n can be seen as an application of an opaque monad transformer Lifting m to a monad n.

Normally, you wouldn't use Lifting in you business code. When generalizing a monad with Reducible, Lifting m n is used as the the most general representation.

Constructors

Lifting

Fields

unwrapLifting :: m a

class MonadLifting m n where

Lifting of actions in monad n into m,

Whereas MonadTrans provides lifting by exactly one transformer layer at a time, MonadLifting can lift by an arbitrary distance at once.

Methods

liftLong :: n a -> m a