Control.Reducible.Orphans

Description

This module contains the orphan instances for some typeclasses that enable them to use Reducible as a fallback,

Orphan instances

(Reducible m, Monad m, MonadFix (Reduced m)) => MonadFix m

Instance details

Methods

mfix :: (a -> m a) -> m a

(Reducible m, Monad m, MonadFail (Reduced m)) => MonadFail m

Instance details

Methods

fail :: String -> m a

(Reducible m, Applicative m, Alternative (Reduced m)) => Alternative m

Instance details

Methods

empty :: m a

(<|>) :: m a -> m a -> m a

some :: m a -> m [a]

many :: m a -> m [a]

(Reducible m, Monad m, MonadPlus (Reduced m)) => MonadPlus m

Instance details

Methods

mzero :: m a

mplus :: m a -> m a -> m a

(Reducible m, Monad m, MonadThrow (Reduced m)) => MonadThrow m

Instance details

Methods

throwM :: Exception e => e -> m a

(Reducible m, Monad m, MonadCatch (Reduced m)) => MonadCatch m

Instance details

Methods

catch :: Exception e => m a -> (e -> m a) -> m a

(Reducible m, Monad m, MonadMask (Reduced m)) => MonadMask m

Instance details

Methods

mask :: ((forall a. m a -> m a) -> m b) -> m b

uninterruptibleMask :: ((forall a. m a -> m a) -> m b) -> m b

generalBracket :: m a -> (a -> ExitCase b -> m c) -> (a -> m b) -> m (b, c)

(Reducible m, Monad m, Monoid w, MonadWriter w (Reduced m)) => MonadWriter w m

Instance details

Methods

writer :: (a, w) -> m a

tell :: w -> m ()

listen :: m a -> m (a, w)

pass :: m (a, w -> w) -> m a

(Reducible m, Monad m, MonadState s (Reduced m)) => MonadState s m

Instance details

Methods

get :: m s

put :: s -> m ()

state :: (s -> (a, s)) -> m a

(Reducible m, Monad m, MonadReader r (Reduced m)) => MonadReader r m

Instance details

Methods

ask :: m r

local :: (r -> r) -> m a -> m a

reader :: (r -> a) -> m a

(Reducible m, Monad m, MonadError e (Reduced m)) => MonadError e m

Instance details

Methods

throwError :: e -> m a

catchError :: m a -> (e -> m a) -> m a